@@ -9,10 +9,9 @@
     "lint": "next lint"
   },
   "dependencies": {
-    "next": "14.2.12",
-    "react": "^18",
-    "react-dom": "^18",
-    "react-router-dom": "^6.26.2"
+    "next": "^14.2.13",
+    "react": "^18.3.1",
+    "react-dom": "^18.3.1"
   },
   "devDependencies": {
     "eslint": "^8",
