@@ -1,9 +1,12 @@
 // src/components/Footer.js
+"use client";
 
 import React from "react";
 
 export default function Footer() {
     return (
-        <p>Footer content goes here</p>
+        <footer>
+            <p class="copyright">&copy; Wiktor Isaksson och Viktor Nordenberg</p>
+        </footer>
     )
 }
