@@ -1,20 +1,23 @@
 // src/components/Header.js
 
 import React from "react";
-import { NavLink } from 'react-router-dom'
+// import { NavLink } from 'react-router-dom'
+import Link from 'next/link';
 
 export default function Header() {
     return (
-        <nav>
-            <ul>
-                <li><Link to="/">Home</Link></li>
-                <li><Link to="/about">About</Link></li>
-                <li><Link to="/contact">Contact</Link></li>
-            </ul>
+        <header>
+            <iframe width="560" height="315" src="https://www.youtube.com/embed/AFUgGZPNT-k?si=9obOo3UyNmLAyf46" title="YouTube video player" frameborder="0" allow="accelerometer; autoplay; clipboard-write; encrypted-media; gyroscope; picture-in-picture; web-share" referrerpolicy="strict-origin-when-cross-origin" allowfullscreen></iframe>
 
-            {/* det kanske ska göras på sättet nedan istället för ovan. Nedan ska vara för SPA, vet inte om ovan är det eller ej */}
-            {/* <NavLink to="/new">Add Event</NavLink> */}
+            {/* <nav>
+                <ul>
+                    <li><Link to="/">Home</Link></li>
+                    <li><Link to="/about">About</Link></li> here is the fcuking problem fix this shit!
+                    <li><Link to="/contact">Contact</Link></li>
+                </ul>
+            </nav> */} 
+        </header>
 
-        </nav>
+
     )
 }
